+++ rk4.cu
@@ -3,10 +3,10 @@
 #include "globalVars.h"
 #include "devFunctionProtos.h"
 
-__device__ void rk4(float *y, float *dydx, int n, float rk4X, float h, float *yout, float iSynap, float ibg, float iff)
+__device__ void rk4(double *y, double *dydx, int n, double rk4X, double h, double *yout, double iSynap, double ibg, double iff)
 {
 	int i;
-	float xh, hh, h6, dym[N_STATEVARS], dyt[N_STATEVARS], yt[N_STATEVARS];
+	double xh, hh, h6, dym[N_STATEVARS], dyt[N_STATEVARS], yt[N_STATEVARS];
 	hh = h*0.5;
 	h6 = h/6.0;
 	xh = rk4X+hh;
